desarrollando clase Libro, Autor, conectando BD

+++ src/main/java/com/proyecto/libreria/dto/LibroDTO.java
@@ -1,7 +1,9 @@
 package com.proyecto.libreria.dto;
 
 public record LibroDTO(Long id,
-                  String titulo,
-                  Double numeroDeDescargas
+                       String titulo,
+                       Long numeroDeDescargas
 ) {
 }
+
+
